201709100656
Small changes.
Because the structure of a web service will not allow relational dbs, I have to make sure the client does not give me crappy data.  This change includes just that.

Change to the customers:
Check to see if the customer has contacts before deleting.  If it does, send back a "forbidden" http response.

Changes to the contacts:
When both PUTting and POSTing, check to see if the customer (parent) does exist.  If it does not, send back a "forbidden" http response.